% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parser.R
\name{.xml_attr}
\alias{.xml_attr}
\title{Get XML attribute}
\usage{
.xml_attr(node, name)
}
\arguments{
\item{node}{XML node}

\item{name}{Attribute name}
}
\value{
Attribute value or NA
}
\description{
Internal function to get an attribute value from an XML node.
}
\keyword{internal}
