% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parser.R
\name{.xml_children}
\alias{.xml_children}
\title{Get XML children}
\usage{
.xml_children(node)
}
\arguments{
\item{node}{XML node}
}
\value{
Child nodes
}
\description{
Internal function to get child nodes of an XML node.
}
\keyword{internal}
