% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parser.R
\name{.xml_name}
\alias{.xml_name}
\title{Get XML node name}
\usage{
.xml_name(node)
}
\arguments{
\item{node}{XML node}
}
\value{
Node name
}
\description{
Internal function to get the name of an XML node.
}
\keyword{internal}
