% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parser.R
\name{.xml_root}
\alias{.xml_root}
\title{Get root element}
\usage{
.xml_root(doc)
}
\arguments{
\item{doc}{XML document}
}
\value{
Root element
}
\description{
Internal function to get the root element of an XML document.
}
\keyword{internal}
