% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_parser.R
\name{.xml_text}
\alias{.xml_text}
\title{Get XML text content}
\usage{
.xml_text(node)
}
\arguments{
\item{node}{XML node}
}
\value{
Text content
}
\description{
Internal function to get text content from an XML node.
}
\keyword{internal}
