% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatograms.R
\name{get_chromatogram}
\alias{get_chromatogram}
\title{Get chromatogram data from mzML file}
\usage{
get_chromatogram(mzml, index)
}
\arguments{
\item{mzml}{Object of class \code{\link[=MzMlFile]{MzMlFile()}}}

\item{index}{Integer chromatogram index (1-based)}
}
\value{
A list containing:
\describe{
\item{time}{Numeric vector of retention times}
\item{intensity}{Numeric vector of intensity values}
\item{id}{Chromatogram identifier}
\item{transition_list}{Transition information if available}
\item{chromatogram_type}{Type of chromatogram (e.g., TIC, BPC)}
}
}
\description{
Extracts time/intensity or other chromatographic data from a single chromatogram.
}
\examples{
\dontrun{
mzml <- read_mzml("path/to/file.mzML")
chrom <- get_chromatogram(mzml, 1)
plot(chrom$time, chrom$intensity, type = "l")
}

}
