% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_file_info}
\alias{get_file_info}
\title{Get file information from mzML object}
\usage{
get_file_info(mzml)
}
\arguments{
\item{mzml}{Object of class \code{\link[=MzMlFile]{MzMlFile()}}}
}
\value{
A list containing:
\describe{
\item{path}{Absolute path to the file}
\item{version}{mzML schema version}
\item{id}{File identifier}
\item{file_content}{Summary of spectrum/chromatogram types in file}
\item{source_files}{List of source RAW files with checksums}
\item{software}{List of software used for processing}
\item{spectrum_count}{Number of spectra in the run}
\item{chromatogram_count}{Number of chromatograms in the run}
\item{start_time}{Acquisition start time (POSIXct)}
\item{ms_levels}{Unique MS levels present in the file}
}
}
\description{
Extracts metadata and summary information from an mzML file using xml2-based
parsing for efficiency.
}
\examples{
\dontrun{
mzml <- read_mzml("path/to/file.mzML")
info <- get_file_info(mzml)
cat("Spectra:", info$spectrum_count, "\n")
cat("MS levels:", paste(info$ms_levels, collapse = ", "), "\n")
}

}
