% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_instrument_info}
\alias{get_instrument_info}
\title{Get instrument configuration information}
\usage{
get_instrument_info(mzml)
}
\arguments{
\item{mzml}{Object of class \code{\link[=MzMlFile]{MzMlFile()}}}
}
\value{
A list containing:
\describe{
\item{instrument_id}{Instrument configuration ID}
\item{components}{List of instrument components (source, analyzer, detector)}
\item{serial_number}{Instrument serial number if available}
\item{manufacturer}{Manufacturer information if available}
}
}
\description{
Extracts instrument hardware and configuration details from an mzML file.
}
\examples{
\dontrun{
mzml <- read_mzml("path/to/file.mzML")
instr <- get_instrument_info(mzml)
print(instr)
}

}
