% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_ms_levels}
\alias{get_ms_levels}
\title{Get MS levels present in the file}
\usage{
get_ms_levels(mzml)
}
\arguments{
\item{mzml}{Object of class \code{\link[=MzMlFile]{MzMlFile()}}}
}
\value{
Integer vector of unique MS levels found
}
\description{
Identifies which mass spectrometry levels (MS1, MS2, etc.) are present.
}
\examples{
\dontrun{
mzml <- read_mzml("path/to/file.mzML")
levels <- get_ms_levels(mzml)
cat("MS levels:", paste(levels, collapse = ", "), "\n")
}

}
